LPS-189701 Move classes in module bean-portlet-registration-impl to module bean-portlet-registration-api, prepare for next
