+++ modules/apps/bean-portlet/bean-portlet-registration-api/src/main/java/com/liferay/bean/portlet/registration/portlet/BeanPortlet.java
@@ -3,10 +3,11 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.bean.portlet.registration.internal;
+package com.liferay.bean.portlet.registration.portlet;
 
 import com.liferay.bean.portlet.extension.BeanPortletMethod;
 import com.liferay.bean.portlet.extension.BeanPortletMethodType;
+import com.liferay.bean.portlet.registration.portlet.app.BeanApp;
 
 import java.util.Dictionary;
 import java.util.List;
