+++ modules/apps/bean-portlet/bean-portlet-registration-api/src/main/java/com/liferay/bean/portlet/registration/portlet/Event.java
@@ -3,17 +3,21 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.bean.portlet.registration.internal;
+package com.liferay.bean.portlet.registration.portlet;
+
+import java.util.List;
 
 import javax.xml.namespace.QName;
 
 /**
  * @author Neil Griffin
  */
-public interface PublicRenderParameter {
+public interface Event {
 
-	public String getIdentifier();
+	public List<QName> getAliasQNames();
 
 	public QName getQName();
+
+	public String getValueType();
 
 }
