+++ modules/apps/bean-portlet/bean-portlet-registration-api/src/main/java/com/liferay/bean/portlet/registration/portlet/app/BeanAppImpl.java
@@ -3,7 +3,10 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.bean.portlet.registration.internal;
+package com.liferay.bean.portlet.registration.portlet.app;
+
+import com.liferay.bean.portlet.registration.portlet.Event;
+import com.liferay.bean.portlet.registration.portlet.PublicRenderParameter;
 
 import java.util.List;
 import java.util.Map;
