+++ modules/apps/bean-portlet/bean-portlet-registration-api/src/main/java/com/liferay/bean/portlet/registration/portlet/util/PortletQNameUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.bean.portlet.registration.internal.xml;
+package com.liferay.bean.portlet.registration.portlet.util;
 
 import com.liferay.portal.kernel.xml.Element;
 import com.liferay.portal.kernel.xml.QName;
